Removes some deprecations after a library upgrade in biz.

+++ src/main/java/sirius/biz/isenguard/RedisLimiter.java
@@ -75,7 +75,7 @@ public class RedisLimiter implements Limiter {
         return getDB().query(() -> "Update rate limit: " + effectiveKey, db -> {
             long value = db.incr(effectiveKey);
             if (value == 1) {
-                db.expire(effectiveKey, intervalInSeconds);
+                db.expire(effectiveKey, (long) intervalInSeconds);
             } else if (value == limit) {
                 limitReachedOnce.run();
             }
